ch4 float_error.c

+++ ch4/float_error.c
@@ -3,8 +3,8 @@
 int main(void)
 {
 	int i;
-	//float f = 0.0;
-	double f = 0.0;
+	float f = 0.0;
+	//double f = 0.0;
 
 	for (i = 0; i < 100; i++)
 		f += 0.1;
